# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estuaire Emploie API Documentation",
        "description": "Documentation de l'API REST pour l'application mobile Estuaire Emploie - Plateforme de recherche d'emploi au Cameroun",
        "contact": {
            "email": "contact@estuaire-emploie.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Liste des entreprises vérifiées",
                "operationId": "a261a49d113a2af2a5ccb5798c6aae7e",
                "responses": {
                    "200": {
                        "description": "Liste des entreprises",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Créer une nouvelle entreprise",
                "description": "Permet à un recruteur de créer le profil de son entreprise avec logo. L'entreprise sera en attente de vérification par l'administrateur.",
                "operationId": "createCompany",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tech Solutions SARL"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contact@techsolutions.cm"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+237 690 123 456"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Entreprise spécialisée dans le développement web et mobile"
                                    },
                                    "logo": {
                                        "description": "Logo de l'entreprise (PNG, JPG, JPEG - max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "sector": {
                                        "type": "string",
                                        "example": "Technologie & IT"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Bonanjo, Rue des Cocotiers"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Douala"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://techsolutions.cm"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entreprise créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entreprise créée avec succès. En attente de vérification."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Détails d'une entreprise",
                "operationId": "50805cb17562834cb99c0477437c5221",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'entreprise",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'entreprise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "jobs": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entreprise non trouvée"
                    }
                }
            }
        },
        "/api/my-company": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Récupérer mon entreprise",
                "description": "Récupère les informations de l'entreprise associée au recruteur connecté",
                "operationId": "getMyCompany",
                "responses": {
                    "200": {
                        "description": "Informations de l'entreprise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "logo": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "sector": {
                                                    "type": "string"
                                                },
                                                "website": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "address": {
                                                    "type": "string"
                                                },
                                                "city": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending",
                                                        "verified",
                                                        "suspended"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Aucune entreprise associée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'avez pas d'entreprise associée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Mettre à jour mon entreprise",
                "description": "Permet au recruteur de mettre à jour les informations de son entreprise avec logo. Utiliser POST au lieu de PUT pour supporter le multipart/form-data.",
                "operationId": "updateMyCompany",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Méthode HTTP (pour Laravel)",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Tech Solutions SARL"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contact@techsolutions.cm"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+237 690 123 456"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "description": "Nouveau logo (PNG, JPG, JPEG - max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "sector": {
                                        "type": "string",
                                        "example": "Technologie & IT"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Douala"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://techsolutions.cm"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entreprise mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entreprise mise à jour avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Non autorisé à modifier cette entreprise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'êtes pas autorisé à modifier cette entreprise"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune entreprise associée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum - Utiliser le token obtenu lors du login/register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}